+++ kernel/user/ubin_registry.c
@@ -60,6 +60,7 @@ extern int utheme_main(int argc, char **argv);
 extern int uwallpaper_main(int argc, char **argv);
 extern int uscale_main(int argc, char **argv);
 extern int uhelp_main(int argc, char **argv);
+extern int uhistory_main(int argc, char **argv);
 extern int uman_main(int argc, char **argv);
 extern int uask_main(int argc, char **argv);
 extern int uaudit_main(int argc, char **argv);